LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.strokeRect.globalcomposite-expected.txt

2d.strokeRect.globalcomposite
strokeRect is not affected by globalCompositeOperation
Actual output:
Expected output:


PASS strokeRect is not affected by globalCompositeOperation
